TEIIDSB-184 creating a common salesforce base
also making a pure interface version of basicconnection
